feat: picking apart monolith go file, adding timestamps to put request
